separate App class
